requirements.txt
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
branca==0.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
Flask==2.3.2
folium==0.14.0
geographiclib==2.0
geopy==2.3.0
idna==3.4
importlib-metadata==6.7.0
ipykernel==6.24.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
newsapi-python==0.2.7
numpy==1.25.0
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.8.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyzmq==25.1.0
requests==2.31.0
six==1.16.0
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
Werkzeug==2.3.6
zipp==3.15.0
